requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.6
django-admin==2.0.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-excel-response2==3.0.6
django-filter==23.3
django-six==1.0.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
excel-base==1.0.4
importlib-metadata==6.8.0
isoweek==1.3.3
kombu==5.3.2
Markdown==3.5
mysqlclient==2.2.0
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
vine==5.0.0
wcwidth==0.2.8
xlwt==1.3.0
zipp==3.17.0
